call

+++ index.js
@@ -3,53 +3,46 @@ const client = new Client({ intents: [GatewayIntentBits.Guilds, GatewayIntentBit
 const json = require('./data.json')
 const ProjectInfo = require('./package.json')
 const fs = require('fs')
-const yt = require('ytdl-core');
+const music = require('./models/Music')
 const commands = require('./commands/commands.js')
 
 client.on('ready', async () => {
     console.log(`Logged in as ${client.user.tag}!`);
     //Adicionar SlashCommands (/)
-    client.application.commands.create(commands.ping)
-    client.application.commands.create(commands.downloadCommand);
-    client.application.commands.create(commands.about)
+    Object.values(commands).forEach((command) => {
+        client.application.commands.create(command)
+    })
 });
 
 //Evento para a deteção dos SlashCommands (/)
 client.on('interactionCreate', async interaction => {
     if (!interaction.isCommand()) return;
   
-    const { commandName, options } = interaction;
+    const { commandName, options } = interaction; 
   
     if (commandName === 'ping') interaction.reply('Pong!');
     if (commandName === 'about') interaction.reply('```json\n' + JSON.stringify(ProjectInfo).toString() + '\n```')
     if (commandName === 'download'){
-        const videoUrl = options.get('url').value
-
-        const audioOptions = {
-            filter: 'audioonly',
-            quality: 'highestaudio'
-        }
-
-        try {
-            await yt.getBasicInfo(videoUrl)
-            interaction.reply("Download started successfully ✅")
-            const audioStream = yt(videoUrl, audioOptions)
-
-            /* Desconmentar caso queira ter feedback de download
-            audioStream.on('progress', (chunkLength, downloaded, total) => {
-                const percent = Math.round((downloaded / total) * 100);
-                console.log(`Download em andamento: ${percent}%`);
-            })
-            */
-            audioStream.on('end', () => {
-                const message = interaction.channel
-                message.send({ files: ['./music/' + interaction.id + '.mp3'] }).then(() => { fs.unlinkSync('./music/' + interaction.id + '.mp3') })
-                .catch((err) => console.log(err))
-            });
-            audioStream.pipe(fs.createWriteStream('./music/' + interaction.id + '.mp3'));
-        } catch (error) {
-            interaction.reply("Err ❌")
-        }
+        music.getmusic(options.get('url').value).then((audioStream) => {
+            if (audioStream != -1){
+                interaction.reply("Download started successfully ✅")
+                audioStream.pipe(fs.createWriteStream('./music/' + interaction.id + '.mp3'));
+                audioStream.on('end', () => {
+                    const message = interaction.channel
+                    message.send({ files: ['./music/' + interaction.id + '.mp3'] }).then(() => { fs.unlinkSync('./music/' + interaction.id + '.mp3') })
+                    .catch((err) => console.log(err))
+                });
+            }
+            else interaction.reply("Err ❌")
+        })
+    }
+    if (commandName === "call"){
+        if (options.get('language') == null) interaction.reply(json['bot-data']['CALL-MSG'] + options.get('game').value + '?')
+        else if (options.get('language').value == 'pt') interaction.reply(json['bot-data']['CALL-MSGPT'] + options.get('game').value + '?')
+        else if (options.get('language').value == 'es') interaction.reply(json['bot-data']['CALL-MSGES'] + options.get('game').value + '?')
+        else if (options.get('language').value == 'fr') interaction.reply(json['bot-data']['CALL-MSGFR'] + options.get('game').value + '?')
+        else if (options.get('language').value == 'it') interaction.reply(json['bot-data']['CALL-MSGIT'] + options.get('game').value + '?')
+        else interaction.reply(json['bot-data']['CALL-MSG'] + options.get('game').value + '?')
     }
 });
